modify type for book price from int to decimal

+++ store/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Book(models.Model):
     title = models.TextField()
     author = models.TextField()
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
 
 
 class CartItem(models.Model):
